Got rid of overload file and updated tests for new API
