+++ tests/kmc_testing.hpp
@@ -422,7 +422,7 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods.",
           "[calc01tot_probs]") {
     double dt = .0001;
     double KBT = 1.;
-    ExampleRod *ep_j[4];
+    std::vector<ExampleRod *> ep_j(4);
     for (int i = 0; i < 4; ++i) {
         ep_j[i] = new ExampleRod;
         *ep_j[i] = MockRod<ExampleRod>(i);
@@ -438,7 +438,6 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods.",
     xlink.LUTablePtr = &LUT;
     // Initialize other vectors for calculations
     std::vector<double> bindFactors(4, xlink.getBindingFactorUS(0));
-    std::vector<int> Uniquefilter(4, 1);
     // Testing
     SECTION("Two semi-overlapping parallel rods with protein in between.") {
         double sep = .25;
@@ -446,24 +445,15 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods.",
         ep_j[0]->pos[1] = 2. * sep;
         ep_j[0]->pos[2] = 0.0;
         double pos[3] = {0, sep, 0};
+
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         xlink.setUnBindPos(pos);
         KMC<ExampleRod> kmc(xlink.getPosPtr(), 2, xlink.getRcutUS(),
                             xlink.getDiffU(), dt);
-        kmc.CalcTotProbsUS(ep_j, Uniquefilter, bindFactors);
-        REQUIRE(kmc.getTotProb() ==
-                Approx(0.000330797337253075).epsilon(SMALL));
-    }
-    SECTION("Test to make sure filter is working.") {
-        // Set filters for rods 2 and 3 to 0
-        Uniquefilter[2] = Uniquefilter[3] = 0;
-        ep_j[0]->pos[0] = 0.0;
-        ep_j[0]->pos[1] = 0.5;
-        ep_j[0]->pos[2] = 0.0;
-        double pos[3] = {0, 0.25, 0};
-        xlink.setUnBindPos(pos);
-        KMC<ExampleRod> kmc(xlink.getPosPtr(), 4, xlink.getRcutUS(),
-                            xlink.getDiffU(), dt);
-        kmc.CalcTotProbsUS(ep_j, Uniquefilter, bindFactors);
+        kmc.CalcTotProbsUS(cep_j, bindFactors);
         REQUIRE(kmc.getTotProb() ==
                 Approx(0.000330797337253075).epsilon(SMALL));
     }
@@ -480,11 +470,15 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods.",
         ep_j[3]->pos[1] = 0.25;
         ep_j[3]->pos[2] = -0.25;
 
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         double pos[3] = {0, 0.25, 0};
         xlink.setUnBindPos(pos);
         KMC<ExampleRod> kmc(xlink.getPosPtr(), 4, xlink.getRcutUS(),
                             xlink.getDiffU(), dt);
-        kmc.CalcTotProbsUS(ep_j, Uniquefilter, bindFactors);
+        kmc.CalcTotProbsUS(cep_j, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.0006615946745061504).epsilon(SMALL));
     }
     // Clean up testing
@@ -499,7 +493,8 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
     double dt = .0001;
     double KBT = 1.;
     // Initialize rods
-    ExampleRod *ep_j[4];
+    std::vector<ExampleRod *> ep_j(4);
+    // ExampleRod *ep_j[4];
     for (int i = 0; i < 4; ++i) {
         ep_j[i] = new ExampleRod;
         *ep_j[i] = MockRod<ExampleRod>(i);
@@ -515,29 +510,35 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
     xlink.LUTablePtr = &LUT;
 
     std::vector<double> bindFactors(4, xlink.getBindingFactorUS(1));
-    std::vector<int> Uniquefilter(4, 1);
 
     SECTION("No rods nearby") {
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
                       ep_j[0]->length, ep_j[0]->rank);
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 1, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, cep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0).margin(SMALL));
     }
     SECTION("One rod nearby and above bound crosslinker") {
         ep_j[1]->pos[0] = 0;
         ep_j[1]->pos[1] = 1.0;
         ep_j[1]->pos[2] = 0;
+
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
                       ep_j[0]->length, ep_j[0]->rank);
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 2, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, cep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.0005481166497130916).epsilon(SMALL));
     }
     SECTION("Two rods nearby, above and below bound crosslinker") {
@@ -549,13 +550,16 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[2]->pos[1] = -1.0;
         ep_j[2]->pos[2] = 0;
 
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
                       ep_j[0]->length, ep_j[0]->rank);
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 3, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, ep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.001096233299426183).epsilon(SMALL));
     }
     SECTION("Two rods nearby, above and perpendicular but same distance from "
@@ -571,14 +575,17 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[2]->direction[1] = 1;
         ep_j[2]->direction[2] = 0;
 
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         // Bind crosslinker to rod
         xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
                       ep_j[0]->length, ep_j[0]->rank);
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 3, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, cep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.001096233299426183).epsilon(SMALL));
     }
     SECTION("Two rods nearby, above and forward touching cutoff.") {
@@ -592,11 +599,14 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[2]->pos[1] = 0;
         ep_j[2]->pos[2] = 0;
         // ep_j[2]->setpos(pos);
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 3, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, cep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(ABS(kmc.getTotProb()) <
               dt * .001); // 3 decimal accuracy more than dt
     }
@@ -612,11 +622,14 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[2]->pos[1] = xlink.posEndBind[0][1] - xlink.getRcutSD() - 1.0;
         ep_j[2]->pos[2] = 0;
 
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 3, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, ep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
 
         CHECK(kmc.getTotProb() == Approx(0).margin(SMALL));
     }
@@ -633,37 +646,19 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[3]->pos[0] = 0;
         ep_j[3]->pos[1] = -1.0;
         ep_j[3]->pos[2] = 0;
+
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
                       ep_j[0]->length, ep_j[0]->rank);
+
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 4, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, ep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.001644349949139275).epsilon(SMALL));
-    }
-    SECTION("Three rods nearby, above and below bound crosslinker. Two "
-            "overlapping but one filtered.") {
-        ep_j[1]->pos[0] = 0;
-        ep_j[1]->pos[1] = 1.0;
-        ep_j[1]->pos[2] = 0;
-
-        ep_j[2]->pos[0] = 0;
-        ep_j[2]->pos[1] = -1.0;
-        ep_j[2]->pos[2] = 0;
-
-        ep_j[3]->pos[0] = 0;
-        ep_j[3]->pos[1] = -1.0;
-        ep_j[3]->pos[2] = 0;
-        Uniquefilter[3] = 0;
-        xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
-                      ep_j[0]->length, ep_j[0]->rank);
-        KMC<ExampleRod> kmc(xlink.posEndBind[0], 4, xlink.getRcutSD(), dt);
-
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
-        CHECK(kmc.getTotProb() == Approx(0.001096233299426183).epsilon(SMALL));
     }
     SECTION("Three rods nearby above, below, and on, all inside rcut.") {
         ep_j[1]->length = ep_j[2]->length = ep_j[3]->length = 2.0;
@@ -675,13 +670,17 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[2]->pos[0] = 0;
         ep_j[2]->pos[1] = -1.0;
         ep_j[2]->pos[2] = 0;
+
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         xlink.setBind(0, ep_j[0]->gid, ep_j[0]->direction, ep_j[0]->pos, 0,
                       ep_j[0]->length, ep_j[0]->rank);
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 4, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, ep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.0003456279751132963).epsilon(SMALL));
     }
     SECTION("Three rods partially inside rcut with different orientations.") {
@@ -703,13 +702,18 @@ TEST_CASE("Test binding probabilities of one head when near multiple rods and "
         ep_j[3]->pos[0] = 0;
         ep_j[3]->pos[1] = 0.5 * ep_j[3]->length + 1.0;
         ep_j[3]->pos[2] = 0.0;
+
+        // Copy ep_j ptr to const ptr vector so it can run with KMC
+        std::vector<const ExampleRod *> cep_j;
+        cep_j.assign(ep_j.begin(), ep_j.end());
+
         KMC<ExampleRod> kmc(xlink.posEndBind[0], 4, xlink.getRcutSD(), dt);
 
-        kmc.CalcTotProbsSD(ep_j, Uniquefilter, ep_j[0]->gid, xlink.lambda,
-                           xlink.kappa, 1. / KBT, xlink.freeLength,
-                           bindFactors);
+        kmc.CalcTotProbsSD(cep_j, ep_j[0]->gid, xlink.lambda, xlink.kappa,
+                           1. / KBT, xlink.freeLength, bindFactors);
         CHECK(kmc.getTotProb() == Approx(0.0007299123928392682).epsilon(SMALL));
     }
+    // Cleanup
     for (int i = 0; i < 4; ++i) {
         delete ep_j[i];
     }
@@ -724,7 +728,7 @@ TEST_CASE("3 crossing perpendicular rods with protein in center",
     // Arrange
     double dt = .0001;
     double KBT = 1.;
-    ExampleRod *ep_j[3];
+    std::vector<ExampleRod *> ep_j(3);
     for (int i = 0; i < 3; ++i) {
         ep_j[i] = new ExampleRod;
         *(ep_j[i]) = MockRod<ExampleRod>(i);
@@ -742,6 +746,10 @@ TEST_CASE("3 crossing perpendicular rods with protein in center",
     ep_j[2]->direction[1] = 0;
     ep_j[2]->direction[2] = 1;
 
+    // Copy ep_j ptr to const ptr vector so it can run with KMC
+    std::vector<const ExampleRod *> cep_j;
+    cep_j.assign(ep_j.begin(), ep_j.end());
+
     ExampleXlink xlink;
     xlink.setMockXlink();
 
@@ -751,14 +759,12 @@ TEST_CASE("3 crossing perpendicular rods with protein in center",
     LookupTable LUT(&lut_filler);
     xlink.LUTablePtr = &LUT;
 
-    std::vector<int> Uniquefilter(3, 1);
-
     // Act
     SECTION("Unbound") {
         KMC<ExampleRod> kmc(xlink.getPosPtr(), 3, xlink.getRcutUS(),
                             xlink.getDiffU(), dt);
         std::vector<double> bindFactorsUS(3, xlink.getBindingFactorUS(0));
-        kmc.CalcTotProbsUS(ep_j, Uniquefilter, bindFactorsUS);
+        kmc.CalcTotProbsUS(cep_j, bindFactorsUS);
         for (int i = 0; i < 3; ++i) {
             // Check individual probabilities
             CHECK(kmc.getProbs(i) ==
@@ -780,7 +786,7 @@ TEST_CASE("3 crossing perpendicular rods with protein in center",
                 double roll = ((j * RodProb) + prevRodProb) / totProb;
                 double calcBindPos = 2. * xlink.rc * (j - .5);
                 double bindpos;
-                int rod_i = kmc.whichRodBindUS(ep_j, bindpos, roll);
+                int rod_i = kmc.whichRodBindUS(cep_j, bindpos, roll);
                 CHECK(rod_i == i);
                 CHECK(bindpos == Approx(calcBindPos).epsilon(SMALL));
             }
@@ -826,9 +832,9 @@ TEST_CASE("3 crossing perpendicular rods with protein in center",
                                 xlink.getRcutSD(), dt);
             std::vector<double> bindFactorsSD(
                 3, xlink.getBindingFactorSD(1 - end_bound));
-            kmc.CalcTotProbsSD(ep_j, Uniquefilter, xlink.idBind[end_bound],
-                               xlink.lambda, xlink.kappa, 1. / KBT,
-                               xlink.freeLength, bindFactorsSD);
+            kmc.CalcTotProbsSD(cep_j, xlink.idBind[end_bound], xlink.lambda,
+                               xlink.kappa, 1. / KBT, xlink.freeLength,
+                               bindFactorsSD);
             // Check individual probabilities
             CHECK(kmc.getProbs(0) == 0.0); // No binding, already attached
             CHECK(kmc.getProbs(1) ==
@@ -881,7 +887,7 @@ TEST_CASE("4 parallel rods separated by a rod diameter on the sides with "
     // Arrange
     double dt = .0001;
     double KBT = 1.;
-    ExampleRod *ep_j[4];
+    std::vector<ExampleRod *> ep_j(4);
     for (int i = 0; i < 4; ++i) {
         ep_j[i] = new ExampleRod;
         *ep_j[i] = MockRod<ExampleRod>(i);
@@ -897,6 +903,11 @@ TEST_CASE("4 parallel rods separated by a rod diameter on the sides with "
     ep_j[1]->direction[0] = -1; // One anti-parallel
     ep_j[2]->direction[0] = 1;
     ep_j[3]->direction[0] = 1;
+
+    // Copy ep_j ptr to const ptr vector so it can run with KMC
+    std::vector<const ExampleRod *> cep_j;
+    cep_j.assign(ep_j.begin(), ep_j.end());
+
     ExampleXlink xlink;
     xlink.setMockXlink();
     // Initialize lookup table
@@ -906,13 +917,12 @@ TEST_CASE("4 parallel rods separated by a rod diameter on the sides with "
     LookupTable LUT(&lut_filler);
     xlink.LUTablePtr = &LUT;
 
-    std::vector<int> Uniquefilter(4, 1);
     // Act
     SECTION("Unbound") {
         KMC<ExampleRod> kmc(xlink.getPosPtr(), 4, xlink.getRcutUS(),
                             xlink.getDiffU(), dt);
         std::vector<double> bindFactorsUS(4, xlink.getBindingFactorUS(0));
-        kmc.CalcTotProbsUS(ep_j, Uniquefilter, bindFactorsUS);
+        kmc.CalcTotProbsUS(cep_j, bindFactorsUS);
         for (int i = 0; i < 4; ++i) {
             // Check individual probabilities
             REQUIRE(kmc.getProbs(i) ==
@@ -934,7 +944,7 @@ TEST_CASE("4 parallel rods separated by a rod diameter on the sides with "
                 double roll = ((j * RodProb) + prevRodProb) / totProb;
                 double calcBindPos = 2. * modRC * (j - .5);
                 double bindpos;
-                int rod_i = kmc.whichRodBindUS(ep_j, bindpos, roll);
+                int rod_i = kmc.whichRodBindUS(cep_j, bindpos, roll);
                 CHECK(rod_i == i);
                 CHECK(bindpos == Approx(calcBindPos).epsilon(SMALL));
             }
@@ -988,9 +998,9 @@ TEST_CASE("4 parallel rods separated by a rod diameter on the sides with "
                                 xlink.getRcutSD(), dt);
             std::vector<double> bindFactorsSD(
                 4, xlink.getBindingFactorSD(1 - end_bound));
-            kmc.CalcTotProbsSD(ep_j, Uniquefilter, xlink.idBind[end_bound],
-                               xlink.lambda, xlink.kappa, 1. / KBT,
-                               xlink.freeLength, bindFactorsSD);
+            kmc.CalcTotProbsSD(cep_j, xlink.idBind[end_bound], xlink.lambda,
+                               xlink.kappa, 1. / KBT, xlink.freeLength,
+                               bindFactorsSD);
             // Check minimum distances
             CHECK(kmc.getDistMin(1) == Approx(.5).epsilon(SMALL));
             CHECK(kmc.getDistMin(2) ==
@@ -1048,7 +1058,8 @@ TEST_CASE("6 perpendicular rods surrounding a sphere of a rod radius with "
     // Arrange
     double dt = .0001;
     double KBT = 1.;
-    ExampleRod *ep_j[6];
+    // ExampleRod *ep_j[6];
+    std::vector<ExampleRod *> ep_j(6);
     for (int i = 0; i < 6; ++i) {
         ep_j[i] = new ExampleRod;
         *ep_j[i] = MockRod<ExampleRod>(i);
@@ -1102,7 +1113,11 @@ TEST_CASE("6 perpendicular rods surrounding a sphere of a rod radius with "
     ep_j[5]->direction[0] = 0.0;
     ep_j[5]->direction[1] = 0.0;
     ep_j[5]->direction[2] = 1.0;
-    std::vector<int> Uniquefilter(6, 1);
+
+    // Copy ep_j ptr to const ptr vector so it can run with KMC
+    std::vector<const ExampleRod *> cep_j;
+    cep_j.assign(ep_j.begin(), ep_j.end());
+
     // Set protein data
     ExampleXlink xlink;
     xlink.setMockXlink();
@@ -1116,7 +1131,7 @@ TEST_CASE("6 perpendicular rods surrounding a sphere of a rod radius with "
         KMC<ExampleRod> kmc(xlink.getPosPtr(), 6, xlink.getRcutUS(),
                             xlink.getDiffU(), dt);
         std::vector<double> bindFactorsUS(6, xlink.getBindingFactorUS(0));
-        kmc.CalcTotProbsUS(ep_j, Uniquefilter, bindFactorsUS);
+        kmc.CalcTotProbsUS(cep_j, bindFactorsUS);
         for (int i = 0; i < 6; ++i) {
             // Check minimum distance
             REQUIRE(kmc.getDistMin(i) == Approx(0.25).epsilon(SMALL));
@@ -1149,7 +1164,7 @@ TEST_CASE("6 perpendicular rods surrounding a sphere of a rod radius with "
                         kmc.getMu(i) + kmc.getDistMin(i) + (modRC * j);
                 }
                 double bindpos;
-                int rod_i = kmc.whichRodBindUS(ep_j, bindpos, roll);
+                int rod_i = kmc.whichRodBindUS(cep_j, bindpos, roll);
                 CHECK(rod_i == i);
                 CHECK(bindpos == Approx(calcBindPos).epsilon(SMALL));
             }
@@ -1203,9 +1218,9 @@ TEST_CASE("6 perpendicular rods surrounding a sphere of a rod radius with "
             std::vector<double> bindFactorsSD(
                 6, xlink.getBindingFactorSD(1 - end_bound));
             // Calculate probabilities
-            kmc.CalcTotProbsSD(ep_j, Uniquefilter, xlink.idBind[end_bound],
-                               xlink.lambda, xlink.kappa, 1. / KBT,
-                               xlink.freeLength, bindFactorsSD);
+            kmc.CalcTotProbsSD(cep_j, xlink.idBind[end_bound], xlink.lambda,
+                               xlink.kappa, 1. / KBT, xlink.freeLength,
+                               bindFactorsSD);
             // Check minimum distances
             CHECK(kmc.getDistMin(1) == Approx(.5).epsilon(SMALL));
             CHECK(kmc.getDistMin(2) ==
@@ -1335,7 +1350,6 @@ TEST_CASE("Works with pointlike objects.") {
     std::vector<const ExampleSphere *> spheres = {&sphere, &sphere2};
     std::vector<const ExampleRod *> rods; // empty vector
     std::vector<double> bind_factors = {0.1, 0.1};
-    std::vector<int> unique_j = {1, 1};
     SECTION("KMC constructed with spheres") {
         KMC<ExampleRod, ExampleSphere> kmc(pos, 0, 1, rc, dt);
 
@@ -1361,8 +1375,8 @@ TEST_CASE("Works with pointlike objects.") {
             bind_fac * 4.0 * M_PI * SQR(sphere.radius) * expon * dt;
         double prob =
             kmc.CalcProbSphereSD(0, sphere, lambda, kappa, beta, l0, bind_fac);
-        kmc.CalcTotProbsSD(rods, spheres, unique_j, 552, lambda, kappa, beta,
-                           l0, bind_factors);
+        kmc.CalcTotProbsSD(rods, spheres, 552, lambda, kappa, beta, l0,
+                           bind_factors);
 
         // Direct calculations
         CHECK(prob == Approx(exp_prob).epsilon(1e-8));
